+++ src/debug.ts
@@ -1,6 +1,5 @@
 import { Keyboard } from "isaac-typescript-definitions";
 import {
-  getRandomSetElement,
   log,
   printConsole,
   saveDataManagerSetGlobal,
@@ -12,8 +11,6 @@ import { hotkeys } from "./modConfigMenu";
 /** Currently, F2 is set to execute this function. */
 function debugCode(_params?: string) {
   // Add code here.
-  const mySet = new Set([{ foo: 1 }, { bar: 2 }, { baz: 3 }]);
-  print(getRandomSetElement(mySet));
 }
 
 /** Hotkey 1 is bound to F2. */