+++ docs/history.md
@@ -3,6 +3,13 @@
 <!-- markdownlint-disable MD001 MD034 -->
 <!-- cspell:ignore Adrayon,aferalsheep,Anidalife,Arsene,Astraii,Bamiot,biriy,Blumpkin,bobita,Bowties,Catgasm,ceehe,Crafterlynx,cutifire,Darkdragon,Darkwolf,Drunkenshoe,Echiqueur,elgirs,fallenfor,Fawkeyes,finalkids,Fyon,Gandi,Ghostbroster,glyndsuresight,goodposter,gorthol,Greninja,Gromfalloon,Ibotep,Iedaim,Inschato,Ivana,Jerseyrebox,Kadda,Karolmo,klover,Krakenos,kwidz,Lbrt,Lobsterosity,makia,Maks,Masterof,meepmeep,mijus,Mirry,missingyes,mithrandi,Moep,molfried,Nanahachiyasu,Nariom,Negi,nicoluwu,Noot,Noowoo,Pachito,Pelo,Pingouin,Thalen,Tehe,Sakk,Satanas,Sheffy,Shigan,sillypears,Sinsthar,sisuka,Smoom,starg,stoogebag,thereisnofuture,Thoday,toooschi,vertopolka,Zazima,Zorua -->
 
+### _v0.85.1_ - August 26th, 2022
+
+- In Season 3, sometimes Dogma will not play its death animation for an unknown reason. If this happens, the screen will immediately turn black (instead of slowly fading out with a static effect) and you will be teleported to the Beast room. If this happens, you will now be automatically teleported back to the Dogma room. (Thanks Cyber_1)
+- Fixed the bug where in season 3 where there would be a trapdoor after the Hush fight. (Thanks TNTerminator)
+- Fixed the bug where in season 3, Schoolbag was awarded twice. (Thanks Cyber_1)
+- Fixed the bug where in season 3, Isaac did not start with the D6 in his active slot.
+
 ### _v0.85.0_ - August 26th, 2022
 
 - In season 3, Tainted Magdalene has been replaced by Isaac.